Move call2vg into a Call2VG class
The class holds all the configuration options and is the authority on default parameters.

Also deduplicates the vcf allele adding functions between call and genotype.
They now live in utility.hpp.
