+++ src/main.cpp
@@ -939,64 +939,20 @@ int main_call(int argc, char** argv) {
     double max_strand_bias = Caller::Default_max_strand_bias;
     string aug_file;
     bool bridge_alts = false;
-    // Option variables (formerly from glenn2vcf)
-    // What's the name of the reference path in the graph?
-    string refPathName = "";
-    // What name should we give the contig in the VCF file?
-    string contigName = "";
-    // What name should we use for the sample in the VCF file?
-    string sampleName = "SAMPLE";
-    // How far should we offset positions of variants?
-    int64_t variantOffset = 0;
-    // How many nodes should we be willing to look at on our path back to the
-    // primary path? Keep in mind we need to look at all valid paths (and all
-    // combinations thereof) until we find a valid pair.
-    int64_t maxDepth = 10;
-    // Shoudl we expect a subgraph and ignore pileups for missing nodes/edges?
+    
+    
+    
+    // Should we expect a subgraph and ignore pileups for missing nodes/edges?
     bool expectSubgraph = false;
-    // Should we write pileup information to the VCF for debugging?
+    
+    // Should we annotate the VCF with pileup info?
     bool pileupAnnotate = false;
-    // What should the total sequence length reported in the VCF header be?
-    int64_t lengthOverride = -1;
-    // What fraction of average coverage should be the minimum to call a variant (or a single copy)?
-    double minFractionForCall = 0;
-    // What fraction of the reads supporting an alt are we willing to discount?
-    // At 2, if twice the reads support one allele as the other, we'll call
-    // homozygous instead of heterozygous. At infinity, every call will be
-    // heterozygous if even one read supports each allele.
-    double maxHetBias = 3;
-    // Like above, but applied to ref / alt ratio (instead of alt / ref)
-    double maxRefHetBias = 4;
-    // How many times more bias do we allow for indels?
-    double indelBiasMultiple = 1;
-    // What's the minimum integer number of reads that must support a call? We
-    // don't necessarily want to call a SNP as het because we have a single
-    // supporting read, even if there are only 10 reads on the site.
-    size_t minTotalSupportForCall = 1;
-    // Bin size used for counting coverage along the reference path.  The
-    // bin coverage is used for computing the probability of an allele
-    // of a certain depth
-    size_t refBinSize = 250;
-    // On some graphs, we can't get the coverage because it's split over
-    // parallel paths.  Allow overriding here
-    size_t expCoverage = 0;
-    // Should we drop variants that would overlap old ones? TODO: we really need
-    // a proper system for accounting for usage of graph material.
-    bool suppress_overlaps = false;
-    // Should we use average support instead minimum support for our calculations?
-    bool useAverageSupport = false;
-    // How big a site should we try to type all at once instead of replacing
-    // with its children if it has any?
-    size_t max_ref_length = 100;
-    // What's the maximum number of bubble path combinations we can explore
-    // while finding one with maximum support?
-    size_t max_bubble_paths = 100;
-    // what's the minimum minimum allele depth to give a PASS in the filter column
-    // (anything below gets FAIL)
-    size_t min_mad_for_filter = 5;
+
+    // This manages conversion from an augmented graph to a VCF, and makes the
+    // actual calls.
+    Call2Vcf call2vcf;
 
     bool show_progress = false;
-    bool verbose = false;
     int thread_count = 1;
 
     int c;
@@ -1074,27 +1030,27 @@ int main_call(int argc, char** argv) {
         // old glenn2vcf opts start here
         case 'r':
             // Set the reference path name
-            refPathName = optarg;
+            call2vcf.refPathName = optarg;
             break;
         case 'c':
             // Set the contig name
-            contigName = optarg;
+            call2vcf.contigName = optarg;
             break;
         case 'S':
             // Set the sample name
-            sampleName = optarg;
+            call2vcf.sampleName = optarg;
             break;
         case 'o':
             // Offset variants
-            variantOffset = std::stoll(optarg);
+            call2vcf.variantOffset = std::stoll(optarg);
             break;
         case 'D':
             // Limit max depth for pathing to primary path
-            maxDepth = std::stoll(optarg);
+            call2vcf.maxDepth = std::stoll(optarg);
             break;
         case 'l':
             // Set a length override
-            lengthOverride = std::stoll(optarg);
+            call2vcf.lengthOverride = std::stoll(optarg);
             break;
         case 'U':
             expectSubgraph = true;
@@ -1104,52 +1060,52 @@ int main_call(int argc, char** argv) {
             break;
         case 'F':
             // Set min fraction of average coverage for a call
-            minFractionForCall = std::stod(optarg);
+            call2vcf.minFractionForCall = std::stod(optarg);
             break;
         case 'H':
             // Set max factor between reads on one alt and reads on the other
             // alt for calling a het.
-            maxHetBias = std::stod(optarg);
+            call2vcf.maxHetBias = std::stod(optarg);
             break;
         case 'R':
             // Set max factor between reads on ref and reads on the other
             // alt for calling a homo ref.
-            maxRefHetBias = std::stod(optarg);
+            call2vcf.maxRefHetBias = std::stod(optarg);
             break;
         case 'M':
             // Set multiplier for bias limits for indels
-            indelBiasMultiple = std::stod(optarg);
+            call2vcf.indelBiasMultiple = std::stod(optarg);
             break;
         case 'n':
             // How many reads need to touch an allele before we are willing to
             // call it?
-            minTotalSupportForCall = std::stoll(optarg);
+            call2vcf.minTotalSupportForCall = std::stoll(optarg);
             break;
         case 'B':
             // Set the reference bin size
-            refBinSize = std::stoll(optarg);
+            call2vcf.refBinSize = std::stoll(optarg);
             break;
         case 'C':
             // Override expected coverage
-            expCoverage = std::stoll(optarg);
+            call2vcf.expCoverage = std::stoll(optarg);
             break;
         case 'O':
             // Suppress variants that overlap others
-            suppress_overlaps = true;
+            call2vcf.suppress_overlaps = true;
             break;
         case 'u':
             // Average (isntead of min) support
-            useAverageSupport = true;
+            call2vcf.useAverageSupport = true;
             break;
         case 'E':
             // Minimum min-allele-depth required to give Filter column a PASS
-            min_mad_for_filter = std::stoi(optarg);
+            call2vcf.min_mad_for_filter = std::stoi(optarg);
             break;
         case 'p':
             show_progress = true;
             break;
         case 'v':
-            verbose = true;
+            call2vcf.verbose = true;
             break;
         case 't':
             thread_count = atoi(optarg);
@@ -1266,30 +1222,10 @@ int main_call(int argc, char** argv) {
     }
 
     // project the augmented graph to a reference path
-    // in order to create a VCF of calls.  this
-    // was once a separate tool called glenn2vcf
-    glenn2vcf::call2vcf(caller._call_graph,
-                        text_file_stream.str(),
-                        refPathName,
-                        contigName,
-                        sampleName,
-                        variantOffset,
-                        maxDepth,
-                        lengthOverride,
-                        pileupAnnotate ? pileup_file_name : string(),
-                        minFractionForCall,
-                        maxHetBias,
-                        maxRefHetBias,
-                        indelBiasMultiple,
-                        minTotalSupportForCall,
-                        refBinSize,
-                        expCoverage,
-                        suppress_overlaps,
-                        useAverageSupport,
-                        max_ref_length,
-                        max_bubble_paths,
-                        min_mad_for_filter,
-                        verbose);
+    // in order to create a VCF of calls.
+    call2vcf.call(caller._call_graph,
+        text_file_stream.str(),
+        pileupAnnotate ? pileup_file_name : string());
 
     return 0;
 }
